refactor: PR comments updated
